requirements.txt
decorator==4.4.0
Flask==1.1.1
Flask-RESTful==0.3.7
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-vision==0.38.1
googleapis-common-protos==1.6.0
graphviz==0.11.1
grpcio==1.22.0
idna==2.8
imageio==2.5.0
imutils==0.5.2
ipykernel==5.1.1
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.0
numpy==1.17.0
opencv-python==4.1.0.25
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
postal==1.1.7
pydotplus==2.0.2
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
pytesseract==0.2.7
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
pyzmq==18.0.2
qtconsole==4.5.2
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1